third_party/blink/web_tests/external/wpt/html/semantics/forms/the-input-element/color-attributes.window-expected.txt

This is a testharness.js-based test.
[FAIL] <input type=color>: alpha attribute
  assert_true: expected true got undefined
[FAIL] <input type=color>: colorspace attribute
  assert_equals: expected (string) "limited-srgb" but got (undefined) undefined
Harness: the test ran to completion.
